Moving the functionality to confirm the cancellation of a pull request review
