+++ src/GitHub.InlineReviews/Services/IPullRequestSessionService.cs
@@ -202,13 +202,6 @@ namespace GitHub.InlineReviews.Services
             ILocalRepositoryModel localRepository,
             string pullRequestId);
 
-
-        /// <summary>
-        /// Displays a confirmation diaglog to ask if the user wants to cancel a pending review.
-        /// </summary>
-        /// <returns></returns>
-        bool ConfirmCancelPendingReview();
-
         /// <summary>
         /// Cancels a pending review on the server.
         /// </summary>
