+++ test/GitHub.InlineReviews.UnitTests/Services/PullRequestSessionTests.cs
@@ -11,7 +11,6 @@ using GitHub.Models;
 using GitHub.Primitives;
 using GitHub.Services;
 using NSubstitute;
-using NSubstitute.Exceptions;
 using NUnit.Framework;
 
 namespace GitHub.InlineReviews.UnitTests.Services
@@ -140,17 +139,14 @@ namespace GitHub.InlineReviews.UnitTests.Services
             }
 
             [Test]
-            public async Task IsFalseWhenReviewCancelledAndConfirmed()
+            public async Task IsFalseWhenReviewCancelled()
             {
                 var currentUser = CreateActor();
                 var review = CreateReview(author: currentUser, state: PullRequestReviewState.Pending);
                 var pr = CreatePullRequest(review);
 
-                var service = Substitute.For<IPullRequestSessionService>();
-                service.ConfirmCancelPendingReview().Returns(true);
-
                 var target = new PullRequestSession(
-                    service,
+                    Substitute.For<IPullRequestSessionService>(),
                     currentUser,
                     pr,
                     Substitute.For<ILocalRepositoryModel>(),
@@ -162,31 +158,6 @@ namespace GitHub.InlineReviews.UnitTests.Services
                 await target.CancelReview();
 
                 Assert.That(target.HasPendingReview, Is.False);
-            }
-
-            [Test]
-            public async Task IsTrueWhenReviewCancelledAndNotConfirmed()
-            {
-                var currentUser = CreateActor();
-                var review = CreateReview(author: currentUser, state: PullRequestReviewState.Pending);
-                var pr = CreatePullRequest(review);
-
-                var service = Substitute.For<IPullRequestSessionService>();
-                service.ConfirmCancelPendingReview().Returns(false);
-
-                var target = new PullRequestSession(
-                    service,
-                    currentUser,
-                    pr,
-                    Substitute.For<ILocalRepositoryModel>(),
-                    "owner",
-                    true);
-
-                Assert.That(target.HasPendingReview, Is.True);
-
-                await target.CancelReview();
-
-                Assert.That(target.HasPendingReview, Is.True);
             }
         }
 
@@ -356,8 +327,6 @@ Line 4";
             public async Task CallsServiceWithNodeId()
             {
                 var service = Substitute.For<IPullRequestSessionService>();
-                service.ConfirmCancelPendingReview().Returns(true);
-
                 var target = CreateTargetWithPendingReview(service);
 
                 await target.CancelReview();
@@ -371,8 +340,6 @@ Line 4";
             public async Task RemovesReviewFromModel()
             {
                 var service = Substitute.For<IPullRequestSessionService>();
-                service.ConfirmCancelPendingReview().Returns(true);
-
                 var target = CreateTargetWithPendingReview(service);
 
                 await target.CancelReview();
